Update patch set 3
Patch Set 3:

It was and this commit is in r23, however they use the open source AOSP compiler, while we use Qualcomm's proprietary compiler, which is supposed to perform better, but since it is closed source, we don't know exactly what they change. Therefore, if they recommend A57 instead of Kryo targeting, it's at the very least worth considering their advice.

Patch-set: 3