Revert "Improve abidump generating logic for non-REL build"
Revert this change to always generate VNDK LSDUMP files.
This is needed for QIIFA VNDK compatibility check.
Current LSdumps needs to be available for comparing against
Golden reference dumps.

As per current logic  VNDK lsdump are generated only if
refernce dumps are present in prebuilts location which is
not inline with QIIFA VNDK plugin architecture.

This reverts commit eb912dcce8b1b7215a9aa47a37ebb8165d7e902a.

CRs-Fixed: 2801394

Change-Id: Iec08ee0c2814cc347cc3f667805c80d797ecd09f