Add support for llndk to vendor snapshot
As part of VSDK, we want to remove framework projects
from the vendor build source tree, including llndk stub
libraries. This change enables llndk stub libraries to
be captured and included at vendor snapshot generation
time, and also to be linked aginst when a vendor build
uses the vendor snapshot.

In the vendor build, when building against the llndk
prebuilt, the build thinks the stub library is simply a
shared library (provided by a vendor_snapshot_shared
module). However, we don't want this library to be
installable, and we *do* want it to be visible to make,
so this change adds the is_llndk proprty. This property
(conveyed from snapshot generation to snapshot
installation via the json IsLlndk proprty) allows Soong
to know that the vendor_snapshot_shared module needs to
have the special treatment not to install, but to be
visible to make.

Bug: 179747111
Bug: 180856097
Change-Id: I9f96d3f6f0a2c65f50060eac3a8748e06287f79c