Use source for hiddenapi if preferred prebuilt is missing boot dex jar
Previously, when a prebuilt was preferred but did not provide a
suitable boot dex jar both the source and the prebuilt were silently
ignored which meant that the "hiddenapi list" command was not given a
complete set of boot dex jars. That could either lead to incorrect
hiddenapi flags being set or the "hiddenapi list" command failing if it
could not find a class.

This change avoids that issue by only using the prebuilt for hiddenapi
if it provides a boot dex jar. Otherwise, hiddenapi will continue to
use the source.

This is only a temporary workaround as using source to generate the
hiddenapi at build time but using the prebuilts at runtime could lead
to inconsistences. Following changes will enable the build to use the
same dex boot jars as the runtime.

Bug: 179459548
Test: python3 vendor/google_keystone/build_qssi.py --build_target=lahaina --use_rbe
Change-Id: Ica4ce42700eb6921e61dd068e31aac37dcd05682